requirements.txt
google-api-python-client==2.97.0  # YouTube Data API
openai>=1.0.0
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.19  # Optional, for more advanced DB interactions
pytest==7.3.1  # For testing
httpx==0.24.1  # HTTP client for OpenAI
typing_extensions==4.7.1
